refactor: encapsulate ddtest code for easy removal
Move all ddtest-specific code into dedicated modules so that removing
ddtest support is a simple matter of deleting a few files and reverting
import lines:

- scripts/ddtest_jobs.py: all ddtest job emission logic (emit_ddtest_jobs,
  validate_ddtest_venv_test_locations, ddtest_k, template helpers) extracted
  from scripts/gen_gitlab_config.py (~200 lines removed from the latter)
- tests/_ddtest_conftest_helpers.py: _DD_PYTEST_XDIST_INFERRED_SERVICE
  pop + cache-seed logic extracted from tests/conftest.py
- tests/_ddtest_env_helpers.py: PYTEST_ADDOPTS stripping extracted from
  tests/utils.py

gen_gitlab_config.py now imports from ddtest_jobs (1 line).
conftest.py now imports from _ddtest_conftest_helpers (1 line).
utils.py now imports from _ddtest_env_helpers (1 line).

To remove ddtest: delete the 3 helper files, revert the 3 import lines,
remove ddtest templates from .gitlab/tests.yml, and remove ddtest: true
from suitespec.yml files.
